Extract FencesBlockToken's message generator out
