+++ src/Microsoft.DocAsCode.Dfm/DfmFencesBlockHelper.cs
@@ -3,12 +3,10 @@
 
 namespace Microsoft.DocAsCode.Dfm
 {
-    using System.Collections.Generic;
-    using System.Collections.Immutable;
-
+    using Microsoft.DocAsCode.Common;
     using Microsoft.DocAsCode.MarkdownLite;
 
-    internal class DfmRendererHelper
+    public static class DfmFencesBlockHelper
     {
         public static string GetRenderedFencesBlockString(DfmFencesBlockToken token, Options options, string errorMessage, string[] codeLines = null)
         {
@@ -34,43 +32,11 @@ namespace Microsoft.DocAsCode.Dfm
             return $"{renderedErrorMessage}{renderedCodeLines}";
         }
 
-        public class SplitToken
+        public static string GenerateReferenceNotFoundErrorMessage(IMarkdownRenderer renderer, DfmFencesBlockToken token)
         {
-            public IMarkdownToken Token { get; set; }
-
-            public List<IMarkdownToken> InnerTokens { get; set; }
-
-            public SplitToken(IMarkdownToken token)
-            {
-                Token = token;
-                InnerTokens = new List<IMarkdownToken>();
-            }
-        }
-
-        public static List<SplitToken> SplitBlockquoteTokens(ImmutableArray<IMarkdownToken> tokens)
-        {
-            var splitTokens = new List<SplitToken>();
-            SplitToken splitToken = null;
-            foreach (var token in tokens)
-            {
-                if (token is DfmSectionBlockToken || token is DfmNoteBlockToken)
-                {
-                    splitToken = new SplitToken(token);
-                    splitTokens.Add(splitToken);
-                }
-                else
-                {
-                    if (splitToken != null)
-                    {
-                        splitToken.InnerTokens.Add(token);
-                        continue;
-                    }
-                    splitToken = new SplitToken(token);
-                    splitToken.InnerTokens.Add(token);
-                    splitTokens.Add(splitToken);
-                }
-            }
-            return splitTokens;
+            string errorMessage = $"Can not find reference {token.Path}";
+            Logger.LogError(errorMessage);
+            return GetRenderedFencesBlockString(token, renderer.Options, errorMessage);
         }
     }
 }
